fixed naming for token details
