+++ src/cashier_frontend/src/components/token-details/hero.tsx
